# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/magic-systems": {
            "get": {
                "description": "Route to get all existing magic systems",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "magic-systems"
                ],
                "summary": "List all magic systems",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MagicSystemGet"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Route to create a new magic system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "magic-systems"
                ],
                "summary": "Create a new magic system",
                "parameters": [
                    {
                        "description": "Magic System Model",
                        "name": "magicSystem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MagicSystemPost"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.MagicSystemPost"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/magic-systems-restore/{id}": {
            "patch": {
                "description": "Route to restore a trashed magic system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "magic-systems"
                ],
                "summary": "Restore a trashed magic system",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Magic System ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/magic-systems-trashed": {
            "get": {
                "description": "Route to get all trashed magic systems",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "magic-systems"
                ],
                "summary": "List all trashed magic systems",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MagicSystemGet"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/magic-systems/{id}": {
            "get": {
                "description": "Route to get a magic system by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "magic-systems"
                ],
                "summary": "Get magic system by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Magic System ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MagicSystemGet"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Route to delete a magic system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "magic-systems"
                ],
                "summary": "Delete a magic system",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Magic System ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Route to update a magic system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "magic-systems"
                ],
                "summary": "Update a magic system",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Magic System ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Magic System Model",
                        "name": "magicSystem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MagicSystemPost"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MagicSystemPost"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shards": {
            "get": {
                "description": "Route to get all existing shards",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shards"
                ],
                "summary": "List all shards",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ShardGet"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Route to create a new shard",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shards"
                ],
                "summary": "Create a new shard",
                "parameters": [
                    {
                        "description": "Shard Model",
                        "name": "shard",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ShardPost"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ShardPost"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shards/trashed": {
            "get": {
                "description": "Route to get all trashed shards",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shards"
                ],
                "summary": "Get trashed shards",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ShardGet"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shards/{id}": {
            "get": {
                "description": "Route to get a shard by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shards"
                ],
                "summary": "Get shard by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Shard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ShardGet"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Route to update a shard",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shards"
                ],
                "summary": "Update a shard",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Shard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Shard Model",
                        "name": "shard",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ShardPost"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ShardPost"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.MagicSystemForShard": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "prerequisit": {
                    "type": "string"
                },
                "shard_id": {
                    "type": "integer"
                }
            }
        },
        "models.MagicSystemGet": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "prerequisit": {
                    "type": "string"
                },
                "shard": {
                    "$ref": "#/definitions/models.ShardForMagicSystem"
                },
                "shard_id": {
                    "type": "integer"
                }
            }
        },
        "models.MagicSystemPost": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "prerequisit": {
                    "type": "string"
                },
                "shard_id": {
                    "type": "integer"
                }
            }
        },
        "models.ShardForMagicSystem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.ShardGet": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "magic_systems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MagicSystemForShard"
                    }
                },
                "name": {
                    "type": "string"
                },
                "planet": {
                    "type": "string"
                },
                "vessel": {
                    "type": "string"
                }
            }
        },
        "models.ShardPost": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "planet": {
                    "type": "string"
                },
                "vessel": {
                    "type": "string"
                }
            }
        }
    }
}